Create _test_monsterStances
some of (or most of) the syntax may be wrong
only other necessary change I can think of is to ensure that Skill cast chance is 0% and apply the classes to the right enemies.
for example, flying enemies like bats and birds should be designated as flying, so as to receive the right stance options
Not familiar with this language. hope this helps a little!